% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignCurves.R
\name{alignCurves}
\alias{alignCurves}
\title{Align the curves}
\usage{
alignCurves(data, days = 1:263, k = 3, n = 5)
}
\description{
\preformatted{@param data Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param k Do you love cats? Defaults to TRUE.
@keywords alignCurves
@examples
}

alignCurves()
@export
}
