% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurve.R
\name{plotCurve}
\alias{plotCurve}
\title{Plot the mean curves for the groups}
\usage{
plotCurve(dataf, days = 1:263, show = "None", min.score = 0, by = "sex")
}
\description{
\preformatted{@param dataf Do you love cats? Defaults to TRUE.
@param days Do you love cats? Defaults to TRUE.
@param show Do you love cats? Defaults to TRUE.
@param min.score Do you love cats? Defaults to TRUE.
@param by Do you love cats? Defaults to TRUE.
@keywords plotCurve
@examples
}

plotCurve()
@export
}
